+++ contracts/Core/storage/VoteStorage.sol
@@ -11,6 +11,6 @@ contract VoteStorage {
     //epoch -> asset -> stakeWeight
     mapping(uint32 => uint256) public totalInfluenceRevealed;
     //epoch -> assetid -> voteValue -> weight
-    mapping(uint32 => mapping(uint8 => mapping(uint256 => uint256))) public voteWeights;
+    mapping(uint32 => mapping(uint8 => mapping(uint32 => uint256))) public voteWeights;
     bytes32 public secrets;
 }
