+++ contracts/IDelegator.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.0;
 interface IDelegator {
     function upgradeDelegate(address newDelegateAddress) external;
 
-    function getResult(uint256 id) external view returns (uint256);
+    function getResult(uint8 id) external view returns (uint32);
 
-    function getJob(uint256 id)
+    function getJob(uint8 id)
         external
         view
         returns (
